requirements.txt
APScheduler==3.6.3
backports.zoneinfo==0.2.1
beautifulsoup4==4.11.1
cachetools==4.2.2
certifi==2022.9.24
charset-normalizer==2.1.1
idna==3.4
lxml==4.9.1
mysql-connector-python==8.0.31
protobuf==3.20.1
pyaes==1.6.1
pyasn1==0.4.8
python-dotenv==0.21.0
python-telegram-bot==13.14
pytz==2022.5
pytz-deprecation-shim==0.1.0.post0
requests==2.28.1
rsa==4.9
six==1.16.0
soupsieve==2.3.2.post1
Telethon==1.25.4
tornado==6.1
tzdata==2022.5
tzlocal==4.2
urllib3==1.26.12
